[bugfix] Port of fix r12943 from trunk.
Latest changes on XMLRPC which involved the usage of QueryResultCache machinery also for resources being fetched did not take into account to update the AbstractQueryResult timestamp each time the a chunk of the resource was fetched. With no updates QueryResultCache can discard those cached resources even when they are still being used. This fix adds a touch() method to AbstractCachedResult, it adds on RpcConnection needed calls to touch() method wherever it is needed and it replaces already existing explicit timestamp attribute updates on AbstractCachedResult instances with calls to touch() method.


git-svn-id: https://exist.svn.sourceforge.net/svnroot/exist/stable/eXist-1.4.x@12944 8784284e-840f-0410-bd5a-a6a6012a0d1a